requirements.txt
certifi==2020.6.20
chardet==3.0.4
ci-info==0.2.0
click==7.1.2
configobj==5.0.6
configparser==5.0.0
decorator==4.4.2
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
httplib2==0.18.1
idna==2.10
isodate==0.6.0
lxml==4.5.2
networkx==2.5
nibabel==3.1.1
nipype==1.5.1
numpy==1.19.1
opencv-python==4.4.0.42
packaging==20.4
pandas==1.1.1
pikepdf==1.19.0
Pillow==7.2.0
prov==1.5.3
pydot==1.4.1
pydotplus==2.0.2
PyMuPDF==1.17.5
pyparsing==2.4.7
PyPDF2==1.26.0
PySide2==5.15.0
pytesseract==0.3.5
python-dateutil==2.8.1
pytz==2020.1
pyxnat==1.3
rdflib==5.0.0
requests==2.24.0
scipy==1.5.2
shiboken2==5.15.0
simplejson==3.17.2
six==1.15.0
traits==6.1.1
urllib3==1.25.10
